create separate test/examples;
